🚀 UPDATE: moved theme folder to styles
