+++ tsconfig.json
@@ -19,7 +19,7 @@
     "baseUrl": ".",
     "paths": {
       "@/components/*": ["src/components/*"],
-      "@/theme/*": ["src/theme/*"],
+      "@/theme/*": ["src/styles/theme/*"],
       "@/styles/*": ["src/styles/*"],
       "@/hooks/*": ["src/hooks/*"],
       "@/config/*": ["src/config/*"],
